Move _utils.ts into crypto_wrappers dir
