+++ src/lib/crypto_wrappers/cms.ts
@@ -1,8 +1,8 @@
 import * as asn1js from 'asn1js';
 import * as pkijs from 'pkijs';
 
-import { deserializeDer, getPkijsCrypto } from '../_utils';
 import * as oids from '../oids';
+import { deserializeDer, getPkijsCrypto } from './_utils';
 import CMSError from './CMSError';
 import Certificate from './x509/Certificate';
